feat: frontend bug fixes, backend enhancements, new modules and UI overhaul

+++ WMS.Application/DTOs/Attendance/AttendanceResponseDto.cs
@@ -7,7 +7,7 @@ namespace WMS.Application.DTOs.Attendance
         public string EmployeeName { get; set; } = string.Empty;
         public DateTime CheckIn { get; set; }
         public DateTime? CheckOut { get; set; }
-        public double? TotalHours { get; set; }
+        public decimal? TotalHours { get; set; }
         public string? WorkMode { get; set; }
         public DateTime AttendanceDate { get; set; }
     }
